# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PropFyndr Backend API",
    "description": "AI-powered real estate advisor backend API. Provides endpoints for chat, property discovery, lead management, analytics, and admin operations.",
    "version": "1.0.0",
    "contact": {
      "name": "PropFyndr Team",
      "email": "team@propfyndr.com"
    },
    "license": {
      "name": "Private"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    },
    {
      "url": "https://api.propfyndr.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/{projectId}": {
      "patch": {
        "tags": [
          "Admin Intelligence"
        ],
        "summary": "Update project intelligence audit status",
        "description": "PATCH endpoint for /api/v1/admin/{projectId}",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/v1/portal/builder/partners": {
      "get": {
        "summary": "Auto-generated endpoint for GET /api/v1/portal/builder/partners",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      },
      "post": {
        "summary": "Auto-generated endpoint for POST /api/v1/portal/builder/partners",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/v1/portal/builder/partners/{id}": {
      "patch": {
        "summary": "Auto-generated endpoint for PATCH /api/v1/portal/builder/partners/:id",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "oneOf": [
              {
                "type": "array"
              },
              {
                "type": "object"
              }
            ]
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token in Authorization header"
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer token from /admin/auth endpoint"
      },
      "WebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "Webhook secret header for lead webhooks"
      }
    }
  }
}